refactor: refactor API server framework
